pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["fitzypop <fitzypop@pm.me>"]

[tool.poetry.dependencies]
python = "3.10.1"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
python-dotenv = "^0.19.2" # neede for pydantic settings management
motor = "^2.5.1" # async mongodb clientg
dnspython = "^2.2.0" # need this package to use `mongodb+srv://` style connections
pydantic = {extras = ["email"], version = "^1.9.0"}
argon2-cffi = "^21.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
devtools = "^0.8.0"
pytest = "^7.1.1"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
